+++ app/src/main/kotlin/io/mszymanski/orknux/server/workflow/WorkflowExceptionResolver.kt
@@ -74,6 +74,7 @@ import io.mszymanski.orknux.server.attachment.AttachmentTooLargeException
 import io.mszymanski.orknux.server.attachment.AttachmentsDisabledException
 import io.mszymanski.orknux.server.attachment.RetentionOutOfRangeException
 import io.mszymanski.orknux.server.trigger.TriggerScheduleInvalidException
+import io.mszymanski.orknux.server.trigger.TriggerScheduleUnreachableException
 import io.mszymanski.orknux.server.variable.VariableCatalogNameInvalidException
 import io.mszymanski.orknux.server.variable.VariableCatalogNameTakenException
 import io.mszymanski.orknux.server.variable.VariableCatalogNotEmptyException
@@ -130,6 +131,7 @@ class WorkflowExceptionResolver : DataFetcherExceptionResolverAdapter() {
             is TriggerConnectionRequiredException,
             is TriggerScheduleRequiredException,
             is TriggerScheduleInvalidException,
+            is TriggerScheduleUnreachableException,
             is TriggerWebhookPathRequiredException,
             is TriggerWebhookPathInvalidException,
             is TriggerWebhookPathTakenException,